% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{append_duckdb}
\alias{append_duckdb}
\title{Append only unique rows to a duckdb table}
\usage{
append_duckdb(x, table, remove_duplicates = TRUE)
}
\arguments{
\item{x}{Object (data frame) to be appended}

\item{table}{Table to append to}

\item{remove_duplicates}{Should duplicate rows be removed? Defaults to TRUE.}
}
\value{
nothing
}
\description{
Append only unique rows to a duckdb table
}
