% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rds.R
\name{append_rds}
\alias{append_rds}
\title{Add data to an existing .rds file}
\usage{
append_rds(x, file, remove_duplicates = TRUE, add_cols = FALSE)
}
\arguments{
\item{x}{Object (data frame) to be appended}

\item{file}{Location of the .rds file}

\item{remove_duplicates}{Should duplicate rows be removed? Defaults to TRUE.}

\item{add_cols}{Should new columns in x be added to the file? If FALSE (default),
will throw an error if new columns are found.}
}
\value{
the full data file (invisibly)
}
\description{
Add data to an existing .rds file
}
