% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_folder}
\alias{check_folder}
\title{Check if a local folder exists, and if not prompt for it.}
\usage{
check_folder(path)
}
\arguments{
\item{path}{the path to check}
}
\value{
the current OR correct path
}
\description{
Check if a local folder exists, and if not prompt for it.
}
