% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{choose_lrh_db}
\alias{choose_lrh_db}
\title{Select and set the LRH Database File}
\usage{
choose_lrh_db()
}
\value{
Nothing
}
\description{
Saves selection in the package environment AND in the Windows Credential
Store for use later
}
