% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{ddb_columns}
\alias{ddb_columns}
\title{Column names and types of a table or view}
\usage{
ddb_columns(x, con = lrh_con())
}
\arguments{
\item{x}{name of the table or view}

\item{con}{connection}
}
\value{
A data.frame with \code{column_name} and \code{column_type}, in position order
}
\description{
Column names and types of a table or view
}
\examples{
\dontrun{
ddb_columns("orders_view")
}
}
