% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{ddb_is_encrypted}
\alias{ddb_is_encrypted}
\title{Check if a ddb file is encrypted}
\usage{
ddb_is_encrypted(db_file)
}
\arguments{
\item{db_file}{File to check}
}
\value{
TRUE or FALSE
}
\description{
Check if a ddb file is encrypted
}
