% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{graph_xl_op}
\alias{graph_xl_op}
\title{Helper function for using the Graph API for coauthoring}
\usage{
graph_xl_op(item, operation, http_verb = "GET", body = NULL)
}
\arguments{
\item{item}{ms_drive_item}

\item{operation}{operation}

\item{http_verb}{verb}

\item{body}{body}
}
\value{
nothing
}
\description{
Helper function for using the Graph API for coauthoring
}
