% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_da2.R
\name{import_da2}
\alias{import_da2}
\title{Import from an \code{XLSX} DA2 Export File}
\usage{
import_da2(
  loc,
  first_header,
  filter_out = c(NA, "Total Cases:"),
  clean_names = TRUE
)
}
\arguments{
\item{loc}{File location}

\item{first_header}{The first header title. Used to identify the starting row}

\item{filter_out}{A character vector of A values to filter out.}

\item{clean_names}{Should column names be cleaned with \code{janitor::clean_names()}?}
}
\value{
a tibble
}
\description{
Import from an \code{XLSX} DA2 Export File
}
