% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_file_locked.R
\name{is_file_locked}
\alias{is_file_locked}
\title{Check if a file is locked for editing}
\usage{
is_file_locked(file)
}
\arguments{
\item{file}{Path to the file to check}
}
\value{
TRUE or FALSE
}
\description{
Check if a file is locked for editing
}
