% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_date.R
\name{last_x}
\alias{last_x}
\alias{last_mo}
\alias{last_q}
\title{Most recent complete month}
\usage{
last_x(unit, flex = lubridate::days(0))

last_mo(flex = lubridate::days(0))

last_q(flex = lubridate::days(0))
}
\arguments{
\item{unit}{What time unit to use. Passed to \link[lubridate:floor_date]{lubridate::floor_date}}

\item{flex}{How flexible to be on the time period? Use \code{days()}, \code{weeks()}, etc.}
}
\value{
a date (no time zone)
}
\description{
Returns the most recent complete time period, given a certain amount of
flex. For example, a \code{flex} value of \code{days(5)} would allow June to count
as the most recent month for June 26th, but not June 25th.
}
