% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{lastupdate_duckdb}
\alias{lastupdate_duckdb}
\alias{lastupdate_alert_duckdb}
\alias{lastupdate_min_duckdb}
\title{Get The time a table as last updated}
\usage{
lastupdate_duckdb(table, dt_col = NULL)

lastupdate_alert_duckdb(table, dt_col = NULL)

lastupdate_min_duckdb(table, min_dt, dt_col = NULL)
}
\arguments{
\item{table}{Table name}

\item{dt_col}{the date column of the table. If not provided, will search for a column that includes "update_dt_tm".}

\item{min_dt}{Generates an error if the last update time is less than this value.}
}
\value{
Prints a cli info statement with the last updated time
}
\description{
\code{lastupdate_duckdb()} returns the raw value of the last update (for use in
scripts). \code{lastupdate_alert_duckdb()} is a convenience function that wraps
the last updated time in a cli-style message.
}
