% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrh_colors.R
\name{lrh_colors}
\alias{lrh_colors}
\alias{show_lrh_colors}
\title{LRH Brand Colors}
\usage{
lrh_colors(...)

show_lrh_colors(
  ...,
  pattern = NULL,
  labels = TRUE,
  borders = NULL,
  cex_label = 1,
  ncol = NULL
)
}
\arguments{
\item{...}{Supply quoted LRH color names to return. If no colors are specified, returns all available colors.}

\item{pattern}{Optional regular expression. If provided, will return only brand colors that match the regular expression}

\item{labels}{Logical. Label colors with names and hex values?}

\item{borders}{Border color for each tile. Default uses \code{par("fg")}. Use \code{border = NA} to omit borders.}

\item{cex_label}{Size of printed labels, as multiplier of default size.}

\item{ncol}{Number of columns. If not supplied, tries to be as square as possible.}
}
\value{
\itemize{
\item \code{lrh_colors()} returns a character vector of color codes
\item \code{show_lrh_colors()} returns nothing
}
}
\description{
Translate human friendly LRH brand color names like "lt_blue" into
accurate hex values for use in plotting. This function can also be used to
show a named vector of all available LRH brand colors and values. Use
\code{show_lrh_colors()} to quickly visualize selected colors in the plot window.
}
\examples{

library(ggplot2)

# Use lrh_colors() to retrieve a single color...
ggplot(mtcars, aes(wt, mpg)) +
  geom_point(color = lrh_colors('green'))

#... multiple colors ...
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  scale_color_manual(values = lrh_colors('green', 'blue', 'red'))

#... or a list of all possible TNTP brand colors
lrh_colors()

# Use show_lrh_colors() to quickly see brand colors in the plotting window
show_lrh_colors('dk_red', 'red', 'lt_red')

# You can also use a pattern to return similar colors
show_lrh_colors(pattern = 'green')

# You can see all colors (and names) by running it with no arguments
show_lrh_colors()
}
