% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{lrh_con}
\alias{lrh_con}
\title{Database Connection Helper}
\usage{
lrh_con(
  db_file = lrh_db(),
  timezone_out = "America/New_York",
  tz_out_convert = "with",
  type = "read_only",
  ...
)
}
\arguments{
\item{db_file}{Location of the DuckDB. Defaults to \code{\link[=lrh_db]{lrh_db()}}}

\item{timezone_out}{Timezone to use for pulling date-time columns}

\item{tz_out_convert}{"with" or "force"}

\item{type}{Either "read_only" or "read_write". Defaults to "read_only"}

\item{...}{Optional additional arguments}
}
\value{
a database connection
}
\description{
Returns a connection to the current duckdb database. Stores the driver and
connection in a package-level environment for use between calls. If calling for
a different type of connection, closes the original connection before opening
the new one and returning it.
}
