% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{lrh_db}
\alias{lrh_db}
\title{LRH Database File Location}
\usage{
lrh_db()
}
\value{
file path to the quality.duckdb file
}
\description{
This function will pull the database file location for your computer.
}
