% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrh_manual.R
\name{lrh_manual_load}
\alias{lrh_manual_load}
\title{Load manual data from an Excel sheet}
\usage{
lrh_manual_load(xl_loc, simplify = TRUE)
}
\arguments{
\item{xl_loc}{Location of the excel file}

\item{simplify}{Should one-column tables be imported as vectors?}
}
\value{
invisibly, the global environment
}
\description{
Each tab will be turned into a tibble and loaded into the Global Environment
with names coming from the sheet names. Sheet names that start with "." will
be skipped (e.g. ".validation" or ".directions")
}
