% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrh_manual.R
\name{lrh_manual_names}
\alias{lrh_manual_names}
\title{List manual data names from an Excel sheet}
\usage{
lrh_manual_names(xl_loc)
}
\arguments{
\item{xl_loc}{Location of the excel file OR an openxlsx2 workbook object}
}
\value{
a character vector of sheet names / variable names
}
\description{
Gives the list of variables that would be imported for a given manual excel
sheet by \link{lrh_manual_load}.
}
