% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{lrh_tableau_folder}
\alias{lrh_tableau_folder}
\title{LRH Tableau Folder}
\usage{
lrh_tableau_folder()
}
\value{
a file path
}
\description{
Helper function to return the local location of the LRH Tableau Folder. Uses
the location of the lrh database (\code{lrh_db()}) to get the location of the
folder
}
