% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{materialize}
\alias{materialize}
\title{Materialize a view}
\usage{
materialize(view, table, quiet = FALSE, materialized_col = FALSE)
}
\arguments{
\item{view}{name of the view to materialize}

\item{table}{name of the table to create}

\item{quiet}{If TRUE doesn't display the success message}

\item{materialized_col}{Add a column \code{materialized} with the current date-time?}
}
\value{
tbl(con, table), invisibly
}
\description{
Materialize a view
}
