% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{materialized_after}
\alias{materialized_after}
\title{Check materialization date}
\usage{
materialized_after(table, dt, dt_col = "materialized", con = lrh_con())
}
\arguments{
\item{table}{table to check}

\item{dt}{date it should be after}

\item{dt_col}{defaults to "materialized"}

\item{con}{connection}
}
\value{
TRUE or FALSE
}
\description{
Check materialization date
}
