% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_check}
\alias{name_check}
\title{Check names for similarities}
\usage{
name_check(names)
}
\arguments{
\item{names}{vector of names to check}
}
\value{
prints the top similarities and silently returns the grid
}
\description{
Check names for similarities
}
