% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_sim}
\alias{name_sim}
\title{Name similarity}
\usage{
name_sim(a, b)
}
\arguments{
\item{a}{vector 1}

\item{b}{vector 2}
}
\value{
a vector of similarities
}
\description{
Get the similarity of two vectors of names after removing common abbreviations
and auto-sorting first vs last names
}
