% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od}
\alias{od}
\title{Get stored OneDrive object. If none is stored, calls od_default() to create
it.}
\usage{
od()
}
\value{
OneDrive object (or NULL)
}
\description{
Get stored OneDrive object. If none is stored, calls od_default() to create
it.
}
