% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_check_folder}
\alias{od_check_folder}
\title{Internal function for checking if a OD folder exists}
\usage{
od_check_folder(od, folder_path)
}
\arguments{
\item{od}{OneDrive (if null, will use the stored OneDrive)}

\item{folder_path}{Folder to check}
}
\value{
nothing. Throws an error if it doesn't exist.
}
\description{
Internal function for checking if a OD folder exists
}
