% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_create_folder}
\alias{od_create_folder}
\title{Create OneDrive Folder}
\usage{
od_create_folder(folder_path, od = NULL)
}
\arguments{
\item{folder_path}{Path to the new folder}

\item{od}{OneDrive (if null, will use the stored OneDrive)}
}
\value{
returns \code{folder_path} invisibly
}
\description{
Create OneDrive Folder
}
