% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_default}
\alias{od_default}
\title{Set default onedrive for od_* functions}
\usage{
od_default(shared = NULL, od = NULL)
}
\arguments{
\item{shared}{Name of shared folder (if setting default to a shared folder)}

\item{od}{OneDrive object to set as default or use for shared folder}
}
\value{
Nothing. Sets default
}
\description{
Set default onedrive for od_* functions
}
