% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_exists}
\alias{od_exists}
\alias{od_locked}
\title{Check if a path / item exists in OneDrive or is locked.}
\usage{
od_exists(path, od = NULL)

od_locked(path, od = NULL)
}
\arguments{
\item{path}{Path to check}

\item{od}{OneDrive (if null, will use the stored OneDrive)}
}
\value{
TRUE or FALSE
}
\description{
Check if a path / item exists in OneDrive or is locked.
}
