% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_get_backup_name}
\alias{od_get_backup_name}
\title{Generate a backup name that doesn't exist}
\usage{
od_get_backup_name(path, od = NULL)
}
\arguments{
\item{path}{path}

\item{od}{OneDrive object or folder object}
}
\value{
a new path
}
\description{
Appends _1 to the file name (or _2 or _3 if needed)
}
