% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_get_shared}
\alias{od_get_shared}
\title{Get an ms_drive object for a shared folder}
\usage{
od_get_shared(name, od = NULL)
}
\arguments{
\item{name}{Name of the shared folder}

\item{od}{OneDrive object (optional)}
}
\value{
an ms_drive object for the shared folder
}
\description{
Get an ms_drive object for a shared folder
}
