% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_get_shortcut}
\alias{od_get_shortcut}
\title{Get an ms_drive_item object for a shortcut folder in your OneDrive}
\usage{
od_get_shortcut(name, od = NULL)
}
\arguments{
\item{name}{Name of shortcut folder}

\item{od}{ONeDrive object (optional)}
}
\value{
an ms_drive_item object for the shortcut folder
}
\description{
Get an ms_drive_item object for a shortcut folder in your OneDrive
}
