% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_list}
\alias{od_list}
\title{List OneDrive Items}
\usage{
od_list(folder = "", od = NULL, pattern = NULL, full_names = FALSE)
}
\arguments{
\item{folder}{Path specification}

\item{od}{OneDrive. If NULL, will use the stored or default onedrive}

\item{pattern}{Optional pattern to filter on.}

\item{full_names}{Should full names be shown?}
}
\value{
tibble of folder contents
}
\description{
List OneDrive Items
}
