% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_read_lines}
\alias{od_read_lines}
\title{Internal function for reading text files from OneDrive}
\usage{
od_read_lines(path, od, ...)
}
\arguments{
\item{path}{path}

\item{od}{OneDrive object or folder object}

\item{...}{additional arguments from od_read()}
}
\value{
data read by readr::read_lines()
}
\description{
Internal function for reading text files from OneDrive
}
