% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_type}
\alias{od_type}
\title{What type of onedrive item is this?}
\usage{
od_type(x = od())
}
\arguments{
\item{x}{}
}
\value{
"ms_drive", "ms_folder", "ms_item" or "unknown"
}
\description{
What type of onedrive item is this?
}
