% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_write_xlsx}
\alias{od_write_xlsx}
\title{Internal funtion for writing xlsx files}
\usage{
od_write_xlsx(x, path, od, ...)
}
\arguments{
\item{x}{item}

\item{path}{path}

\item{od}{onedrive}

\item{...}{additional parameters}
}
\value{
nothing
}
\description{
Internal funtion for writing xlsx files
}
