% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_da2.R
\name{parse_dt_tm}
\alias{parse_dt_tm}
\title{Internal parse date-time function}
\usage{
parse_dt_tm(x, tz)
}
\arguments{
\item{x}{character vector}

\item{tz}{Timezone to use when converting date/time columns}
}
\value{
a date-time or date column
}
\description{
Internal parse date-time function
}
