% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ph_with.gt_tbl}
\alias{ph_with.gt_tbl}
\title{For adding gt tables to slides}
\usage{
\method{ph_with}{gt_tbl}(x, value, location, ...)
}
\arguments{
\item{x}{an rpptx object}

\item{value}{gt() object}

\item{location}{location. See [officer::ph_with()]}

\item{...}{Additional arguments passed on to external image ph_with()}
}
\value{
x
}
\description{
Uses chrome to print the table to an image, then adds it using
[officer::ph_with.external_img()]. By default, makes the table as large as
possible for the location without changing its scale and centers it.
}
