% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt_as_dml}
\alias{ppt_as_dml}
\title{Convert ggplot to editable vector graphic}
\usage{
ppt_as_dml(content)
}
\arguments{
\item{content}{content to potentially convert}
}
\value{
content, possibly wrapped in [rvg::dml()]
}
\description{
If the content is a ggplot object, wraps it with [rvg::dml()] so it is
inserted as an editable vector graphic. Otherwise returns the content as-is.
}
\keyword{internal}
