% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt_d2}
\alias{ppt_d2}
\title{Add a double data slide}
\usage{
ppt_d2(p, left, right, notes = NULL)
}
\arguments{
\item{p}{rpptx object (from \code{\link[=ppt_lrh]{ppt_lrh()}})}

\item{left}{content to add to left pane}

\item{right}{content to add to right pane}

\item{notes}{optional speaker notes for the slide}
}
\value{
rpptx object
}
\description{
Add a double data slide
}
