% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt_save}
\alias{ppt_save}
\title{Save Powerpoint file}
\usage{
ppt_save(p, location = NULL, open = FALSE)
}
\arguments{
\item{p}{file to save}

\item{location}{location. If omitted, will open a dialog to select a location}

\item{open}{Should the file be opened after saving?}
}
\value{
p (invisibly)
}
\description{
Save Powerpoint file
}
