% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_date.R
\name{prep_date}
\alias{prep_date}
\title{Prepend a formatted date on the file name}
\usage{
prep_date(file, date, fmt = "\%Y-\%m-\%d", sep = ".")
}
\arguments{
\item{file}{File name}

\item{date}{Date reference}

\item{fmt}{Formatting string. See \link{strptime} for details}

\item{sep}{Separator}
}
\value{
a formatted file name with the date prepended followed by "."
}
\description{
Prepend a formatted date on the file name
}
