% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_date.R
\name{prep_ymd}
\alias{prep_ymd}
\alias{prep_ym}
\alias{prep_y}
\title{Prepend a formatted date on the file name}
\usage{
prep_ymd(file, date, fmt = "\%Y-\%m-\%d", sep = ".")

prep_ym(file, date, fmt = "\%Y-\%m", sep = ".")

prep_y(file, date, fmt = "\%Y", sep = ".")
}
\arguments{
\item{file}{File name}

\item{date}{Date reference}

\item{fmt}{Formatting string. See \link{strptime} for details}

\item{sep}{Separator between date and filename}
}
\value{
a formatted file name with the date prepended followed by "."
}
\description{
Prepends a date to the file part of a file location string. So the location
\code{foo/bar.csv} becomes \verb{foo/2025-01-01.bar.csv}. \code{prep_ymd()} uses a default
of YYYY-MM-DD. \code{prep_ym()} uses a default of YYYY-MM, and \code{prep_y()} uses YYYY
}
