% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{pull_duckdb}
\alias{pull_duckdb}
\title{Quickly pull a table from Duck DB}
\usage{
pull_duckdb(table, db_file = lrh_db, con = NULL)
}
\arguments{
\item{table}{Name of table}

\item{db_file}{Location of the duckdb file (defaults to [lrh_db])}

\item{con}{(Optional) connection object if you are already connected to the DB.}
}
\value{
a tibble
}
\description{
Convenience function for pulling a single table. Handles connecting and
disconnecting. For more advanced queries, create the connection once and
use existing `DBI` functions.
}
