% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lrh.R
\name{register_lrh_fonts}
\alias{register_lrh_fonts}
\title{Register fonts with system fonts}
\usage{
register_lrh_fonts(
  fonts = c("Open Sans SemiCondensed ExtraBold", "Open Sans SemiCondensed SemiBold")
)
}
\arguments{
\item{fonts}{defaults to LRH theme fonts}
}
\value{
Nothing
}
\description{
Needed for using \code{\link[officer:ph_with]{officer::ph_with()}} with ggplot objects. Runs on package
load automatically.
}
