% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_flatten}
\alias{sql_flatten}
\title{Flattening function for use in dbplyr piplines}
\usage{
sql_flatten(x, sep = "; ", distinct = TRUE, order_by = NULL)
}
\arguments{
\item{x}{values to aggregate}

\item{sep}{Separator. Defaults to "; "}

\item{distinct}{Should values be made distinct?}

\item{order_by}{values to order by. Defaults to \code{x}}
}
\value{
a \code{sql()} STRING_AGG statement
}
\description{
Use with the \verb{!!} operator to return the sql statement before translation
}
