% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau_auth}
\alias{tableau_auth}
\title{Authenticate in Tableau}
\usage{
tableau_auth(
  base_url = "https://us-east-1.online.tableau.com/api/3.15",
  site_content_url = "lrh_nh",
  pat_name = "RStudio",
  pat_secret = tntpr::tntp_cred("TABLEAU_PAT")
)
}
\arguments{
\item{base_url}{Tableau Base URL}

\item{site_content_url}{Site name}

\item{pat_name}{PAT Name}

\item{pat_secret}{PAT}
}
\value{
nothing
}
\description{
Authenticates in tableau and stores the credentials token for use with other
\verb{tableau_*} functions. By default, credentials tokens are valid for 120 minutes
on Tableau Cloud -- if using for longer you will need to reauthorize.
}
\details{
Defaults are currently set for LRH, but can be adjusted
for any site. For more details see \href{https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_ref_authentication.htm#sign_in}{Tableau's REST Documentation}
}
