% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tableau_data_path}
\alias{tableau_data_path}
\title{Path helper for the Sharepoint Tableau Data Folder}
\usage{
tableau_data_path(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[=file.path]{file.path()}}}
}
\value{
a file path
}
\description{
Helper function to return the local location of the LRH Tableau Data Folder.
Can be run with arguments like \code{\link[=file.path]{file.path()}}
}
