% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau_get_datasource_id}
\alias{tableau_get_datasource_id}
\title{Get data source ID by name}
\usage{
tableau_get_datasource_id(auth, datasource_name, project_name = NULL)
}
\arguments{
\item{auth}{Authentication list from tableau_sign_in()}

\item{datasource_name}{Name of the data source}

\item{project_name}{Optional project name to filter by (recommended for disambiguation)}
}
\value{
Data source ID string
}
\description{
Get data source ID by name
}
