% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau_get_project_id}
\alias{tableau_get_project_id}
\title{Get project ID by name}
\usage{
tableau_get_project_id(auth, project_name)
}
\arguments{
\item{auth}{Authentication list from tableau_sign_in()}

\item{project_name}{Name of the project/folder}
}
\value{
Project ID string
}
\description{
Get project ID by name
}
