% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau_list_data}
\alias{tableau_list_data}
\title{List Tableau Data sources}
\usage{
tableau_list_data()
}
\value{
a tibble with details on all current Tableau data sources
}
\description{
List Tableau Data sources
}
