% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{upsert_duckdb}
\alias{upsert_duckdb}
\title{Upsert in DuckDB}
\usage{
upsert_duckdb(table, x, id_col, dt_col)
}
\arguments{
\item{table}{Table name}

\item{x}{data frame to update}

\item{id_col}{the id column of the table}

\item{dt_col}{the date column of the table}
}
\value{
number of rows affected
}
\description{
Performs an upsert on a duckdb table. If the table doesn't yet exist, this
function will prompt the user if they would like to create it.
}
