% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{write_tz_duckdb}
\alias{write_tz_duckdb}
\title{Write a dataframe to duckdb using TIMESTAMPTZ for date-time fields}
\usage{
write_tz_duckdb(table, x, overwrite = FALSE)
}
\arguments{
\item{table}{Name of table to write to}

\item{x}{dataframe}

\item{overwrite}{Overwrite a current table (if it exists)?}
}
\value{
x (invisibly)
}
\description{
Write a dataframe to duckdb using TIMESTAMPTZ for date-time fields
}
