% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{alias_check}
\alias{alias_check}
\title{Check names against current aliases}
\usage{
alias_check(
  names = character(0),
  table = "PG_PROVIDER_ALIAS",
  sensitivity = 0.7,
  print = 10
)
}
\arguments{
\item{names}{character vector of names to check}

\item{table}{table to check names against}

\item{sensitivity}{How similar do names need to be to trigger an audit? Set
to 0 to always audit.}

\item{print}{How many rows of the table should be printed? Set to 0 or FALSE to hide}
}
\value{
the name check data frame, invisibly
}
\description{
Check names against current aliases
}
