% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{append_duckdb}
\alias{append_duckdb}
\title{Append only unique rows to a duckdb table}
\usage{
append_duckdb(
  x,
  table,
  db_file = lrh_db(),
  remove_duplicates = TRUE,
  con = NULL
)
}
\arguments{
\item{x}{Object (data frame) to be appended}

\item{table}{Table to append to}

\item{db_file}{Location of the duckdb file (defaults to \code{\link[=lrh_db]{lrh_db()}})}

\item{remove_duplicates}{Should duplicate rows be removed? Defaults to TRUE.}

\item{con}{(Optional) connection object if you are already connected to the DB.}
}
\value{
nothing
}
\description{
Append only unique rows to a duckdb table
}
