% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{connect_duckdb}
\alias{connect_duckdb}
\title{Helper function to connect to the LRH DuckDB}
\usage{
connect_duckdb(db_file = lrh_db, ...)
}
\arguments{
\item{db_file}{Location of the DuckDB. Defaults to [lrh_db]}

\item{...}{Optional additional arguments}
}
\value{
a database connection
}
\description{
Helper function to connect to the LRH DuckDB
}
