% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{graph_df_to_values}
\alias{graph_df_to_values}
\title{Convert an R data frame to the nested list format the Graph API expects}
\usage{
graph_df_to_values(df)
}
\arguments{
\item{df}{data frame to convert}
}
\value{
a nested list
}
\description{
Convert an R data frame to the nested list format the Graph API expects
}
