% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_da2.R
\name{import_da2}
\alias{import_da2}
\title{Import from an DA2 Query CSV}
\usage{
import_da2(
  file,
  guess_max = Inf,
  ddb_reference = NULL,
  tz = "America/New_York",
  ...
)
}
\arguments{
\item{file}{file location}

\item{guess_max}{from \code{readr::read_csv()}. Defaults to infinite (slightly
slower, but no errors)}

\item{ddb_reference}{Optional: A duckdb table name to use as a reference for
the input. This ensures columns are read with the same column types as the
duckdb table. Uses \code{pull_duckdb()}}

\item{tz}{Timezone to use when converting date/time columns}

\item{...}{Other arguments passed on to \code{readr::read_csv()}}
}
\value{
a tibble
}
\description{
Convenience function. Imports the csv (assuming infinite guessing), cleans
column names and converts any timestamp columns
}
