% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrh_csv.R
\name{lrh_csv}
\alias{lrh_csv}
\title{LRH CSV writer}
\usage{
lrh_csv(x, file, na = "", row.names = FALSE, ...)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not,
it is attempted to coerce x to a data frame.}

\item{file}{either a character string naming a file or a connection open for
writing. "" indicates output to the console.}

\item{na}{the string to use for missing values in the data.}

\item{row.names}{either a logical value indicating whether the row names of x
are to be written along with x, or a character vector of row names to be written.}

\item{...}{Additional arguments passed on to \code{write.csv()}}
}
\value{
\code{x}, invisibly
}
\description{
Wraps \code{write.csv()} with some standard defaults. Unlike \code{readr::write_csv()},
this function will write date-times in their current timezone (rather than
translating to UTC).
}
