% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\docType{data}
\name{lrh_db}
\alias{lrh_db}
\title{Quality Database Location}
\format{
An object of class \code{character} of length 1.
}
\usage{
lrh_db
}
\description{
Hard-coded location of the quality database on the network.
}
\keyword{datasets}
