% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrh_csv.R
\name{lrh_excel}
\alias{lrh_excel}
\title{LRH Excel Writer}
\usage{
lrh_excel(
  x,
  widths = "auto",
  table_style = "TableStyleMedium1",
  file = NULL,
  open = is.null(file)
)
}
\arguments{
\item{x}{a dataframe or list of dataframes}

\item{widths}{column widths (default: auto)}

\item{table_style}{Excel table style (default: gray/striped)}

\item{file}{path to write to (defaults to tempfile)}

\item{open}{open the file after writing? By default, opens tempfiles but not
other files}
}
\value{
the path, invisibly
}
\description{
Uses \code{openxlsx2} to write to an excel file. Accepts a single data frame or a
list of data frames for multiple tabs. Data is written as a table, and column
widths are set to "auto" by default.
}
\details{
By default, writes to a tempfile and opens the file, but can also be used to
write to specified location with the \code{file} parameter
}
