% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{materialize}
\alias{materialize}
\title{Materialize a view}
\usage{
materialize(view, table, con = lrh_con(), quiet = FALSE)
}
\arguments{
\item{view}{name of the view to materialize}

\item{table}{name of the table to create}

\item{con}{Connection to use. Defaults to \code{\link[=lrh_con]{lrh_con()}}}

\item{quiet}{If TRUE doesn't display the success message}
}
\value{
tbl(con, table), invisibly
}
\description{
Materialize a view
}
