% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_default}
\alias{od_default}
\title{Set default onedrive for od_* functions}
\usage{
od_default(
  folder = NULL,
  shortcut = NULL,
  shared = NULL,
  itemid = NULL,
  od = NULL
)
}
\arguments{
\item{folder}{A folder within the provided OneDrive or shared folder to set as the default.}

\item{shortcut}{The name of a folder that has been added as a shortcut to your OneDrive}

\item{shared}{Name of shared folder (if setting default to a shared folder)}

\item{itemid}{A specific folder id (instead of a folder name)}

\item{od}{OneDrive object to set as default or use for shared folder}
}
\value{
the OneDrive object (invisibly)
}
\description{
Set default onedrive for od_* functions
}
