% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_list}
\alias{od_list}
\title{List OneDrive Items}
\usage{
od_list(
  folder = "",
  od = NULL,
  pattern = NULL,
  full_names = FALSE,
  recursive = FALSE,
  info = "partial"
)
}
\arguments{
\item{folder}{Path specification}

\item{od}{OneDrive. If NULL, will use the stored or default onedrive}

\item{pattern}{Optional pattern to filter on.}

\item{full_names}{Should full names be shown?}

\item{recursive}{If TRUE, recursively list all files in subfolders.
When TRUE, full_names is forced to TRUE and only files (not folders)
are returned. Default is FALSE.}

\item{info}{"name", "partial", or "all". "partial" and "all" return data.frames,
"name" returns a vector of just names}
}
\value{
tibble of folder contents
}
\description{
List OneDrive Items
}
