% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_read}
\alias{od_read}
\alias{od_write}
\title{Read/Write from OneDrive}
\usage{
od_read(path, od = NULL, type = NULL, ...)

od_write(x, path, od = NULL, ...)
}
\arguments{
\item{path}{The location in the Sharepoint drive}

\item{od}{OneDrive (if null, will use the stored OneDrive)}

\item{type}{Optional. One of "dataframe" (for delimited files), "xlsx", or "rds". Uses the file extension to determine type if not provided.}

\item{...}{Additional arguments passed on to the reading/writing function.}

\item{x}{file to save}
}
\value{
\code{od_read()} returns an R object as specified by type. \code{od_write()}
returns x, invisibly
}
\description{
Based off of \code{\link[tntpr:sp_read]{tntpr::sp_read()}} and \code{\link[tntpr:sp_read]{tntpr::sp_write()}}.

Read or write data to/from a OneDrive Folder. Can be used with default
folder/drive set by \code{\link[=od_default]{od_default()}} or with a specified folder/drive.

Currently supported file types include: \code{.csv}, \code{.csv2}, \code{.tsv}, \code{.xls},
\code{.xlsx}, \code{.rds}

These functions will attempt to use the appropriate read/write function based
on the file extension, however this can be overridden by specifying type.

The \code{...} parameter is passed on to the appropriate reading or writing
function. See the details section for more information on these functions
by type.

If the folder in \code{path} does not yet exist, the user will be prompted if they
would like to create it.
}
\section{Details}{
For more information on methods (shown as \verb{$__()} below) see documentation
on \link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive}.
\subsection{Reading Functions}{
\itemize{
\item ".csv", ".csv2", ".tsv" are read using the \verb{$load_dataframe()} method,
which uses \code{\link[readr:read_delim]{readr::read_delim()}}.
\item ".rds" is read using the \verb{$load_rds()} method which accepts no additional
arguments.
\item ".xls" and ".xlsx" are read using \code{\link[readxl:read_excel]{readxl::read_excel()}} (if installed).
The function will download the excel file temporarily, then import it and
delete the temporary copy
}
}

\subsection{Writing Functions}{
\itemize{
\item ".csv", ".csv2", ".tsv" are written using the \verb{$save_dataframe()} method
and uses \code{\link[readr:write_delim]{readr::write_delim()}}. Delimiter will be assumed by the extension
unless provided in a \code{delim} argument
\item ".rds" is written using the \verb{$save_rds()} method, which accepts no
additional arguments
\item ".xlsx" is written using \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}} (if
installed) and then uploaded using the \verb{$upload_file()} method.
}
}
}

\seealso{
\code{\link[=od_upload]{od_upload()}}, \code{\link[=od_download]{od_download()}}; \verb{$upload_file()}, \verb{$download_file()}, \verb{$save_rdata()}, \verb{$load_rdata()} from \link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive}
}
