% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_xl_compare}
\alias{od_xl_compare}
\title{Compare a dataframe to an existing spreadsheet}
\usage{
od_xl_compare(x, path, table, id_cols, od = NULL, wb_types = NULL)
}
\arguments{
\item{x}{Data frame of rows to append (columns must match the table)}

\item{path}{The location in the Sharepoint drive}

\item{table}{Name of the Excel Table}

\item{id_cols}{Column name (or vector of names) to use as an id.}

\item{od}{OneDrive (if null, will use the stored OneDrive)}

\item{wb_types}{Used for [openxlsx2::wb_to_df()] `types` parameter.}
}
\value{
a list of (append, patch) for use with od_xl_append() and od_xl_patch()
}
\description{
Returns a list of (append, patch)
}
