% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{od_xl_patch}
\alias{od_xl_patch}
\title{Patch cells through the coauthoring API}
\usage{
od_xl_patch(x, path, od = NULL)
}
\arguments{
\item{x}{Data frame with columns \code{sheet}, \code{range}, and \code{new}}

\item{path}{The location in the Sharepoint drive}

\item{od}{OneDrive (if null, will use the stored OneDrive)}
}
\value{
the ms_drive_item (invisibly)
}
\description{
This is designed to work with \code{od_xl_compare()}, but can also be done
manually by creating a data frame with columns \code{sheet}, \code{range}, and \code{new}.
}
