% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt_lrh}
\alias{ppt_lrh}
\title{Create new Powerpoint}
\usage{
ppt_lrh(
  title = NULL,
  subtitle = NULL,
  template_loc = "~/Custom Office Templates/LRH Data PPT Template.potx"
)
}
\arguments{
\item{title}{Title Text (optional)}

\item{subtitle}{Subtitle Text (optional)}

\item{template_loc}{Location of PPT Template}
}
\value{
a rpptx object from \link{officer}
}
\description{
Extends \link{officer} to create a new LRH Data PPT Slide with a title and subtitle
}
