% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{pull_duckdb}
\alias{pull_duckdb}
\title{Quickly pull a table from Duck DB}
\usage{
pull_duckdb(table, max_rows = Inf)
}
\arguments{
\item{table}{Name of table}

\item{max_rows}{Maximum number of rows to pull (by default, pulls all)}
}
\value{
a tibble
}
\description{
Convenience function for pulling a single table. Handles connecting and
disconnecting. For more advanced queries, create the connection once and
use existing `DBI` functions.
}
