% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{replace_duckdb}
\alias{replace_duckdb}
\title{Replace rows in a table}
\usage{
replace_duckdb(x, table, delete_where, db_file = lrh_db, con = NULL)
}
\arguments{
\item{x}{Object (data frame) to be appended}

\item{table}{Table name}

\item{delete_where}{String of a SQL Where selection for rows to delete}

\item{db_file}{Location of the duckdb file (defaults to [lrh_db])}

\item{con}{(Optional) connection object if you are already connected to the DB.}
}
\value{
Nothing
}
\description{
This convenience function first deletes rows according to the where statement
and then appends the rows from x
}
