% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_age}
\alias{sql_age}
\alias{age}
\title{Age at a given time}
\usage{
sql_age(dt, dob)

age(dt, dob)
}
\arguments{
\item{dt}{date at which to find the age}

\item{dob}{date of birth}
}
\value{
\code{sql_age()} returns a sql() statement, \code{age()} returns an integer age
}
\description{
\code{sql_age()} is designed to be used in lazy tibbles. Use with the \verb{!!}
operator for local evaluation into a \code{sql()} string first.
}
