% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_as_date}
\alias{sql_as_date}
\title{As Date with TZ function for dbplyr pipelines}
\usage{
sql_as_date(x, tz = "America/New_York")
}
\arguments{
\item{x}{values to cast to date}

\item{tz}{timezone to cast at.}
}
\value{
a \code{sql()} CAST statement
}
\description{
Use with the \verb{!!} operator to return the sql statement before translation
}
