% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_ce_dt_tm}
\alias{sql_ce_dt_tm}
\alias{sql_ce_date}
\title{Process DA-2 Date Strings}
\usage{
sql_ce_dt_tm(x, tz = "America/New_York")

sql_ce_date(x)
}
\arguments{
\item{x}{column with date string (ce_result_value)}

\item{tz}{Timezone (for date-times)}
}
\value{
a \code{sql()} CAST statement
}
\description{
These strings take the form of '0:2025112718300000:0.000000:126:0'
}
