% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lrh.R
\name{theme_lrh}
\alias{theme_lrh}
\title{LRH ggplot2 theme}
\usage{
theme_lrh(
  grid = FALSE,
  grid_color = "#CBCBCB",
  legend_style = "top",
  border = FALSE,
  md = FALSE,
  base_font = "Open Sans",
  title_font = "Open Sans SemiCondensed ExtraBold",
  subtitle_font = "Open Sans SemiCondensed SemiBold",
  ...
)
}
\arguments{
\item{grid}{Which grid lines should be shown? Use \code{TRUE} or \code{FALSE} to toggle
all grid lines, or a string combination of \code{X}, \code{x}, \code{Y}, \code{y} for major and minor x and y grid lines.}

\item{grid_color}{Color of grid lines added by the \code{grid} parameter. Defaults to \code{"#CBCBCB"}.}

\item{legend_style}{Either "top" or "right"}

\item{border}{Include a panel border?}

\item{md}{Should text elements use \code{\link[ggtext:element_markdown]{ggtext::element_markdown()}}?}

\item{base_font}{Basic font}

\item{title_font}{Title font}

\item{subtitle_font}{Subtitle font}

\item{...}{Additional arguments passed on to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
a ggplot2 theme statement
}
\description{
Cobbled together with inspiration from \code{\link[tntpr:tntp_style]{tntpr::tntp_style()}} and the default
\code{ggplot2} themes.
}
