% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{write_tz_duckdb}
\alias{write_tz_duckdb}
\title{Write a dataframe to duckdb using TIMESTAMPTZ for date-time fields}
\usage{
write_tz_duckdb(table, x, con = NULL, db_file = lrh_db())
}
\arguments{
\item{table}{Name of table to write to}

\item{x}{dataframe}

\item{con}{(Optional) connection object if you are already connected to the DB.}

\item{db_file}{Location of the duckdb file (defaults to \code{\link[=lrh_db]{lrh_db()}})}
}
\value{
x (invisibly)
}
\description{
Write a dataframe to duckdb using TIMESTAMPTZ for date-time fields
}
