% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_.R
\name{xl_unprotect}
\alias{xl_unprotect}
\alias{xl_protect}
\title{Unprotect sheets, run an expression, then re-protect}
\usage{
xl_unprotect(item, sheets)

xl_protect(item, sheets)
}
\arguments{
\item{item}{An ms_drive_item for the workbook}

\item{sheets}{Character vector of sheet names to unprotect}
}
\value{
NULL (called for side effects)
}
\description{
Sheets are always re-protected on exit, even if the expression errors.
}
